fix: improve avatar loading failure detection.
Avatar load failure detection was subject to a race condition between the browser fetching the image and the sites JS. In many cases, the image was loaded before unpoly could even initialize meaning the callback system would not see a failure since the image never loaded. Now we do a check to see if the naturalHeight is 0 and assume the image failed if it is.